+++ src/components/button/stories/Button.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/svelte";
-import { Button } from ".";
+import { Button } from "../";
 import ButtonStory from "./Button.story.svelte";
 
 // More on how to set up stories at: https://storybook.js.org/docs/7.0/svelte/writing-stories/introduction
