+++ src/components/input/Input.svelte
@@ -12,14 +12,19 @@
   export let label = "";
   export let id: string;
   export let error = "";
+
+  let hasError = false;
+
+  $: hasError = error != null && error.length > 0;
 </script>
 
-<label for={id} class="text-background text-sm">
+<label for={id} class="text-background text-sm" class:text-danger={hasError}>
   {label}
   <div class="relative">
     <input
       {id}
-      class="bg-gray-50 border border-gray-300 text-background text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 disabled:bg-gray-200"
+      class="bg-gray-50 border border-gray-300 text-background text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 disabled:bg-gray-100"
+      class:border-danger={hasError}
       {...$$restProps}
       bind:value
       on:input
@@ -28,7 +33,7 @@
       <slot name="sufix" />
     </div>
   </div>
-  {#if error != null && error.length > 0}
+  {#if hasError}
     <legend class="text-danger text-sm mt-1">{error}</legend>
   {/if}
 </label>
