+++ src/components/wallet-providers/stories/WalletProviders.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/svelte";
-import { WalletProviders } from ".";
+import { WalletProviders } from "../";
 
 // More on how to set up stories at: https://storybook.js.org/docs/7.0/svelte/writing-stories/introduction
 const meta = {
